+++ Frontend/src/routes/dashboard/+page.server.ts
@@ -2,19 +2,26 @@ import type {PageServerLoad} from './$types';
 
 export const load: PageServerLoad = () => {
   return {
+    keywords: [
+      {label: 'Sharlene', value: 'DE'},
+      {label: 'Peter', value: 'FR'},
+      {label: 'Troy', value: 'US'},
+    ],
     identities: [
       {
-        picture: 'https://randomuser.me/api/portraits/women/27.jpg',
+        id: 'dd2c249985de',
+        avatar: 'https://randomuser.me/api/portraits/women/27.jpg',
         name: 'Sharlene Olson',
         country: 'DE',
-        location: 'Stuttgart, Baden-Württemberg',
+        location: 'Stuttgart, baden',
         email: 'sharlene@google.com',
         card: 8293_5562_1822_7463,
         wallet: '16cZ4Qvd13ZuTGN6CsMJhJLUjoetvRDdPU',
         phone: 156_806_7120,
       },
       {
-        picture: 'https://randomuser.me/api/portraits/men/33.jpg',
+        id: 'acf7a18abfd8',
+        avatar: 'https://randomuser.me/api/portraits/men/33.jpg',
         name: 'Peter Ferguson',
         country: 'FR',
         location: 'Paris, Ile-de-France',
@@ -24,7 +31,8 @@ export const load: PageServerLoad = () => {
         wallet: 'xqcZ4Q4513ZuTGzd5sMJhJLzjoetvRDdnE',
       },
       {
-        picture: 'https://randomuser.me/api/portraits/men/1.jpg',
+        id: '18129fe791a1',
+        avatar: 'https://randomuser.me/api/portraits/men/1.jpg',
         name: 'Troy Cruz',
         country: 'US',
         location: 'San Francisco, CA',