+++ Frontend/src/lib/icons/Accounts.svelte
@@ -7,6 +7,7 @@
   viewBox="0 0 24 24"
   stroke-width="2"
   class={className + ' small-icon fill-none stroke-current'}>
-  <path stroke="none" d="M0 0h24v24H0z" fill="none" /><path
-    d="M5 4h4l2 5l-2.5 1.5a11 11 0 0 0 5 5l1.5 -2.5l5 2v4a2 2 0 0 1 -2 2a16 16 0 0 1 -15 -15a2 2 0 0 1 2 -2" />
+  <path stroke="none" d="M0 0h24v24H0z" fill="none" /><path d="M9 7m-4 0a4 4 0 1 0 8 0a4 4 0 1 0 -8 0" /><path
+    d="M3 21v-2a4 4 0 0 1 4 -4h4a4 4 0 0 1 4 4v2" /><path d="M16 3.13a4 4 0 0 1 0 7.75" /><path
+    d="M21 21v-2a4 4 0 0 0 -3 -3.85" />
 </svg>