+++ CHANGELOG.md
@@ -5,6 +5,19 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.3.02] - 2024-12-6
+
+### Added
+
+- Add search functionality
+- Add filter/sort functionality
+- Add new icons
+
+## Changed
+
+- Modify dashboard-table responsiveness
+- Change table columns
+
 ## [0.3.01] - 2024-12-4
 
 ### Added